Implement QueryRuntime
